% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranfast.R
\name{moranfast}
\alias{moranfast}
\title{Calculate Moran's I quickly for point data}
\usage{
moranfast(x, c1, c2, alternative = "two.sided")
}
\arguments{
\item{x}{a numeric vector, the value distributed over space}

\item{c1}{a numeric vector, the first (x) value of a column of coordinates}

\item{c2}{a numeric vector, the second (y) value of a column of coordinates}

\item{alternative}{a character sring specifying the alternative hypothesis that is tested against; must be one of "two.sided", "less", or "greater", or any unambiguous abbreviation of these.}
}
\description{
Calculate Moran's I quickly for point data
}
